+++ src/App.js
@@ -70,6 +70,9 @@ const Payment = React.lazy(() => import("./pages/Payment"));
 const Affiliate = React.lazy(() => import("./pages/Affiliate"));
 const Support = React.lazy(() => import("./pages/Support"));
 const Privacy = React.lazy(() => import("./pages/PrivacyPolicy"));
+const TermsofUSe = React.lazy(() => import("./pages/TermsofUse"));
+const AMLPolicy = React.lazy(() => import("./pages/AMLPolicy"));
+const CookiePolicy = React.lazy(() => import("./pages/CookiePolicy"));
 const Imprint = React.lazy(() => import("./pages/Imprint"));
 const PasswordForgot = React.lazy(() => import("./pages/PasswordForgot"));
 const Product = React.lazy(() => import("./pages/Product"));
@@ -426,6 +429,36 @@ function App() {
                             <Navigation />
 
                             <Support />
+                            <Footer />
+                          </>
+                        }
+                      />
+                      <Route
+                        path="/termsofuse"
+                        element={
+                          <>
+                            <Navigation />
+                            <TermsofUSe />
+                            <Footer />
+                          </>
+                        }
+                      />
+                      <Route
+                        path="/aml-policies"
+                        element={
+                          <>
+                            <Navigation />
+                            <AMLPolicy />
+                            <Footer />
+                          </>
+                        }
+                      />
+                      <Route
+                        path="/cookie-policies"
+                        element={
+                          <>
+                            <Navigation />
+                            <CookiePolicy />
                             <Footer />
                           </>
                         }